feat: move 3rd party source code into separate folder
