--- src/styles/components/_MessageTabBar.js
@@ -1,25 +1,0 @@
-import { StyleSheet } from 'react-native';
-import colors from '../common/_colors';
-
-export default StyleSheet.create({
-  tab: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    paddingBottom: 10,
-    flexDirection: 'row',
-  },
-  tabs: {
-    height: 50,
-    flexDirection: 'row',
-    justifyContent: 'space-around',
-    borderWidth: 1,
-    borderTopWidth: 0,
-    borderLeftWidth: 0,
-    borderRightWidth: 0,
-    borderColor: '#ccc',
-  },
-  alert: {
-    marginLeft: -5,
-  },
-});
